requirements.txt
Flask==2.1.1
tensorflow==2.12.0
numpy==1.24.2
opencv-python-headless==4.8.0.74
Pillow==9.4.0  # For image processing
